Changes to fix type mismatch.

+++ fractint-float/common/encoder.c
@@ -525,7 +525,7 @@ int encoder()
           osave_info.oymax     = oymax;
           osave_info.ox3rd     = ox3rd;
           osave_info.oy3rd     = oy3rd;
-          osave_info.keep_scrn_coords= keep_scrn_coords;
+          osave_info.keep_scrn_coords= (short)keep_scrn_coords;
           osave_info.drawmode  = drawmode;
           for (i = 0; i < sizeof(osave_info.future) / sizeof(short); i++)
              osave_info.future[i] = 0;
@@ -810,7 +810,7 @@ static void _fastcall setup_save_info(struct fractal_info far * save_info)
    save_info->closeprox = closeprox;
    save_info->nobof = (short) nobof;
    save_info->orbit_interval = orbit_interval;
-   save_info->orbit_delay = orbit_delay;
+   save_info->orbit_delay = (short) orbit_delay;
    save_info->math_tol[0] = math_tol[0];
    save_info->math_tol[1] = math_tol[1];
    for (i = 0; i < sizeof(save_info->future) / sizeof(short); i++)
